@@ -1,6 +1,6 @@
 module github.com/variantdev/chartify
 
-go 1.17
+go 1.20
 
 require (
 	github.com/davecgh/go-spew v1.1.1
